requirements.txt
web3==5.31.4
py-solc-x==1.1.1
pytest==7.4.3
